+++ src/proof_system/widget/permutation.rs
@@ -10,4 +10,4 @@ mod verifierkey;
 
 pub(crate) use proverkey::ProverKey;
 
-pub(crate) use verifierkey::VerifierKey;
+pub(crate) use verifierkey::VerificationKey;